+++ CHANGELOG.md
@@ -6,6 +6,7 @@ master
 
 * Fix: nil pointer dereference when Network was not found
 * Update hcloud-go to 1.18.1
+* Add `HCLOUD_LOAD_BALANCERS_ENABLED` env variable to disable the Load Balancer feature, per default the Load Balancers are enabled
 
 v1.6.1
 ------